backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
sqlparse==0.5.3
psycopg2-binary
django-cors-headers
djangorestframework
django_extensions

python-dotenv
pytest
requests

mongoengine
pymongo